reorganize plot commands, fix t-sne
