+++ devtools/conda-recipe/meta.yaml
@@ -32,7 +32,7 @@ test:
     - gpy
     - msmbuilder
     - coverage
-    - python-coveralls
+    - bokeh
 
   imports:
     - osprey
